starting up a 6 layer sac2_0 but with the gpt2ish_med heads and embedding size to see how it stacks against gpt2ish_med
